[#2] Refactor: type 및 constants 관련 중복제거 및 상수 추가
orderList에서 중복되는 SortType과 StatusType을 삭제하고 PAGINATION_PER_PAGE와 TODAY의 위치를 변경했습니다.
units의 이름을 orderData로 변경했습니다.
