+++ src/components/OrderTable/OrderTableBody.tsx
@@ -2,7 +2,8 @@ import TableBody from '@mui/material/TableBody';
 import TableCell from '@mui/material/TableCell';
 import TableRow from '@mui/material/TableRow';
 
-import { PAGINATION_PER_PAGE } from '@/api/orderList';
+import { PAGINATION_PER_PAGE } from '@/constants';
+
 import { OrderProps, SortType, StatusType } from '@/types/props';
 
 interface OrderTableBodyProps {
