+++ src/constants/orderData.ts
@@ -1,4 +1,7 @@
 export const ITEMS_PER_PAGE = 50;
+export const PAGINATION_PER_PAGE = 50;
+export const TODAY = '2023-03-08';
+export const DEFAULT_PAGE = '1';
 
 export const orderApiQueryKey = {
   PAGE: 'page',
